refactor: change folder structure and refactor message cards
